requirements.txt
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
Flask-Login==0.4.0